@@ -6,25 +6,26 @@ authors = [
     { name = "Mickey Kim", email = "mickey.kim@gmail.com" }
 ]
 dependencies = [
-    "fastapi>=0.109.0",
+    "fastapi>=0.115.0",
     "uvicorn[standard]>=0.27.0",
     "sqlalchemy>=2.0.25",
     "alembic>=1.13.1",
     "psycopg2-binary>=2.9.9",
-    "pydantic[email]>=2.5.3",
-    "pydantic-settings>=2.1.0",
-    "python-jose[cryptography]>=3.3.0",
-    "passlib[bcrypt]>=1.7.4",
+    "pydantic[email]>=2.10.3",
+    "pydantic-settings>=2.7.0",
     "python-multipart>=0.0.6",
     "boto3>=1.34.15",
     "requests>=2.31.0",
-    "mangum>=0.17.0",
     "python-dotenv>=1.0.0",
     "tenacity>=8.2.3",
     "email-validator>=2.1.0.post1",
     "bcrypt>=4.2.1",
+    "structlog>=24.4.0",
+    "sqlalchemy-utils>=0.41.2",
+    "pyjwt>=2.10.1",
+    "cryptography>=44.0.0",
 ]
-requires-python = ">=3.10"
+requires-python = ">=3.12"
 
 [dependency-groups]
 dev = [
@@ -41,6 +42,9 @@ dev = [
     "types-requests>=2.31.0",
     "sqlalchemy-utils>=0.41.1",
     "pytest-mock>=3.12.0",
+    "bandit>=1.8.0",
+    "pip-audit>=2.7.3",
+    "pylint>=3.3.2",
 ]
 
 [build-system]
@@ -55,7 +59,7 @@ allow-direct-references = true
 
 [tool.black]
 line-length = 88
-target-version = ['py310']
+target-version = ['py312']
 include = '\.pyi?$'
 
 [tool.isort]
@@ -68,7 +72,7 @@ line_length = 88
 
 [tool.ruff]
 line-length = 88
-target-version = "py310"
+target-version = "py312"
 lint.select = [
     "E",  # pycodestyle errors
     "W",  # pycodestyle warnings
@@ -79,7 +83,8 @@ lint.select = [
 ]
 exclude = [
     ".venv",
-    "alembic/"
+    "alembic/",
+    "init-db.py"
 ]
 
 [tool.mypy]
@@ -101,6 +106,7 @@ module = [
   "botocore.exceptions.*",
   "jose.*",
   "passlib.context",
+  "sqlalchemy_utils.*"
 ]
 
 [tool.pytest.ini_options]
@@ -116,4 +122,8 @@ env = [
     "POSTGRES_PORT=5432",
     "POSTGRES_DB=test_db",
     "SECRET_KEY=test_secret_key",
+    "AWS_ACCESS_KEY_ID=test",
+    "AWS_SECRET_ACCESS_KEY=test",
+    "AWS_REGION=eu-west-2",
 ]
+asyncio_default_fixture_loop_scope = "function"
