@@ -1,9 +1,9 @@
 from typing import Annotated, Generator
 from uuid import UUID
 
+import jwt
 from fastapi import Depends, HTTPException, status
 from fastapi.security import OAuth2PasswordBearer
-from jose import JWTError, jwt
 from sqlalchemy.orm import Session
 
 from app.core.config import get_settings
@@ -11,8 +11,8 @@ from app.core.database import SessionLocal
 from app.schemas.user import UserInDB
 from app.services.user_service import UserService
 
-settings = get_settings()
-oauth2_scheme = OAuth2PasswordBearer(tokenUrl=f"{settings.API_V1_STR}/auth/login")
+SETTINGS = get_settings()
+OAUTH2_SCHEME = OAuth2PasswordBearer(tokenUrl=f"{SETTINGS.API_V1_STR}/auth/login")
 
 
 def get_db() -> Generator[Session, None, None]:
@@ -24,7 +24,7 @@ def get_db() -> Generator[Session, None, None]:
 
 
 async def get_current_user(
-    token: Annotated[str, Depends(oauth2_scheme)],
+    token: Annotated[str, Depends(OAUTH2_SCHEME)],
     db: Annotated[Session, Depends(get_db)],
 ) -> UserInDB:
     credentials_exception = HTTPException(
@@ -33,12 +33,14 @@ async def get_current_user(
         headers={"WWW-Authenticate": "Bearer"},
     )
     try:
-        payload = jwt.decode(token, settings.SECRET_KEY, algorithms=["HS256"])
+        payload = jwt.decode(token, SETTINGS.SECRET_KEY, algorithms=["HS256"])
         user_id: str = payload.get("sub")
         if user_id is None:
             raise credentials_exception
-    except JWTError as err:
-        raise credentials_exception from err
+    except jwt.ExpiredSignatureError as e:
+        raise credentials_exception from e
+    except jwt.PyJWTError as e:
+        raise credentials_exception from e
 
     user_service = UserService(db)
     user = user_service.get_by_id(UUID(user_id))
