@@ -1,4 +1,8 @@
-from sqlalchemy import create_engine
+"""
+Database methods
+"""
+
+from sqlalchemy import Engine, create_engine
 from sqlalchemy.orm import DeclarativeBase, Session, sessionmaker
 
 from app.core.config import get_settings
@@ -9,10 +13,8 @@ settings = get_settings()
 class Base(DeclarativeBase):
     """Base class for all SQLAlchemy models."""
 
-    pass
 
-
-engine = create_engine(
+_engine = create_engine(
     settings.DATABASE_URL,
     pool_pre_ping=True,
     pool_size=5,
@@ -20,5 +22,10 @@ engine = create_engine(
 )
 
 SessionLocal = sessionmaker(
-    autocommit=False, autoflush=False, bind=engine, class_=Session
+    autocommit=False, autoflush=False, bind=_engine, class_=Session
 )
+
+
+def get_engine() -> Engine:
+    """Returns engine for SQLAlchemy models."""
+    return _engine
