@@ -1,3 +1,7 @@
+"""
+Base API for bank interactions
+"""
+
 from abc import ABC, abstractmethod
 from datetime import datetime
 from typing import Optional, Type
@@ -6,6 +10,8 @@ from typing_extensions import TypedDict
 
 
 class TokenResponse(TypedDict):
+    """Bank token response"""
+
     access_token: str
     refresh_token: Optional[str]
     expires_at: datetime
@@ -14,13 +20,9 @@ class TokenResponse(TypedDict):
 class BankAPIError(Exception):
     """Base exception for bank API errors."""
 
-    pass
-
 
 class TokenRefreshError(BankAPIError):
     """Raised when token refresh fails."""
-
-    pass
 
 
 class BankAPI(ABC):
@@ -29,18 +31,16 @@ class BankAPI(ABC):
     @abstractmethod
     def refresh_token(self, refresh_token: str) -> TokenResponse:
         """Refresh access token using refresh token."""
-        pass
 
     @abstractmethod
     def exchange_code(self, auth_code: str) -> TokenResponse:
         """Exchange authorization code for access token."""
-        pass
 
 
 def get_bank_api(account_type: str) -> BankAPI:
     """Factory function to get appropriate bank API implementation."""
     # Import here to avoid circular imports
-    from app.services.bank_api.monzo import MonzoAPI
+    from app.services.bank_api.monzo import MonzoAPI  # pylint: disable=cyclic-import
 
     apis: dict[str, Type[BankAPI]] = {
         "monzo": MonzoAPI,
