@@ -3,6 +3,13 @@ import { useNavigate } from 'react-router-dom';
 import { AlertCircle } from 'lucide-react';
 import { useAuth } from '@/context/AuthContext';
 
+interface ValidationError {
+  type: string;
+  loc: string[];
+  msg: string;
+  input: string | null;
+}
+
 export const LoginForm = () => {
   const [email, setEmail] = useState('');
   const [password, setPassword] = useState('');
@@ -17,29 +24,23 @@ export const LoginForm = () => {
     setError('');
 
     try {
-      const response = await fetch('http://localhost:8000/api/v1/auth/login', {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/x-www-form-urlencoded',
-        },
-        body: new URLSearchParams({
-          username: email,
-          password: password,
-        }),
-      });
-
-      const data = await response.json();
-
-      if (!response.ok) {
-        throw new Error(data.detail);
-      }
-
-      if (data.access_token) {
-        await login(data.access_token);
-        navigate('/dashboard');
-      }
+      await login(email, password);
+      navigate('/dashboard');
     } catch (err: any) {
-      setError(err.message);
+      console.error('Login form error:', err);
+      if (err.response?.status === 422) {
+        const validationErrors = err.response.data.detail;
+        if (Array.isArray(validationErrors)) {
+          // Format validation errors nicely
+          setError(validationErrors.map((error: ValidationError) => error.msg).join(', '));
+        } else {
+          setError('Invalid input. Please check your credentials.');
+        }
+      } else if (err.response?.status === 401) {
+        setError(err.response.data.detail || 'Incorrect email or password');
+      } else {
+        setError('Failed to login. Please try again.');
+      }
     } finally {
       setIsLoading(false);
     }
@@ -47,21 +48,25 @@ export const LoginForm = () => {
 
   return (
     <div className="min-h-screen bg-gray-50 flex flex-col justify-center py-12 sm:px-6 lg:px-8">
+      <div className="sm:mx-auto sm:w-full sm:max-w-md">
+        <h2 className="mt-6 text-center text-3xl font-extrabold text-gray-900">
+          Sign in to your account
+        </h2>
+      </div>
+
       <div className="mt-8 sm:mx-auto sm:w-full sm:max-w-md">
         <div className="bg-white py-8 px-4 shadow sm:rounded-lg sm:px-10">
           {error && (
-            <div className="mb-4 p-4 text-red-700 bg-red-100 rounded-md">
-              <div className="flex">
-                <AlertCircle className="h-5 w-5 mr-2" />
-                <span>{error}</span>
-              </div>
+            <div className="mb-4 p-4 text-red-700 bg-red-100 rounded-md flex items-center">
+              <AlertCircle className="h-5 w-5 mr-2 flex-shrink-0" />
+              <span className="text-sm">{error}</span>
             </div>
           )}
           
           <form className="space-y-6" onSubmit={handleSubmit}>
             <div>
               <label htmlFor="email" className="block text-sm font-medium text-gray-700">
-                Email address
+                Email
               </label>
               <input
                 id="email"
@@ -71,8 +76,7 @@ export const LoginForm = () => {
                 required
                 value={email}
                 onChange={(e) => setEmail(e.target.value)}
-                className="mt-1 block w-full rounded-md border-gray-300 shadow-sm"
-                disabled={isLoading}
+                className="mt-1 block w-full rounded-md border-gray-300 shadow-sm focus:border-indigo-500 focus:ring-indigo-500 sm:text-sm"
               />
             </div>
 
@@ -88,15 +92,14 @@ export const LoginForm = () => {
                 required
                 value={password}
                 onChange={(e) => setPassword(e.target.value)}
-                className="mt-1 block w-full rounded-md border-gray-300 shadow-sm"
-                disabled={isLoading}
+                className="mt-1 block w-full rounded-md border-gray-300 shadow-sm focus:border-indigo-500 focus:ring-indigo-500 sm:text-sm"
               />
             </div>
 
             <button
               type="submit"
               disabled={isLoading}
-              className="w-full flex justify-center py-2 px-4 border border-transparent rounded-md shadow-sm text-sm font-medium text-white bg-indigo-600 hover:bg-indigo-700 disabled:opacity-50"
+              className="w-full flex justify-center py-2 px-4 border border-transparent rounded-md shadow-sm text-sm font-medium text-white bg-indigo-600 hover:bg-indigo-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500 disabled:opacity-50"
             >
               {isLoading ? 'Signing in...' : 'Sign in'}
             </button>
@@ -106,5 +109,3 @@ export const LoginForm = () => {
     </div>
   );
 };
-
-export default LoginForm;
