@@ -1,27 +1,26 @@
+// src/test/mocks/handlers.ts
 import { http, HttpResponse } from 'msw';
+import { mockUser, mockBankAccount } from '../utils';
 
 export const handlers = [
-  // Successful login
-  http.post('/api/v1/auth/login', async ({ request }) => {
-    const formData = await request.text();
-    if (formData.includes('test@example.com')) {
-      return HttpResponse.json({
-        access_token: 'fake-token',
-        token_type: 'bearer'
-      });
-    }
-    return HttpResponse.json(
-      { detail: 'Incorrect email or password' },
-      { status: 401 }
-    );
+  // Simplified login handler
+  http.post('/api/v1/auth/login', async () => {
+    return HttpResponse.json({
+      access_token: 'fake-token',
+      token_type: 'bearer'
+    });
   }),
 
+  // Simplified user info handler
   http.get('/api/v1/auth/me', () => {
+    return HttpResponse.json(mockUser);
+  }),
+
+  // Simplified bank accounts handler
+  http.get('/api/v1/bank-accounts', () => {
     return HttpResponse.json({
-      id: '123',
-      email: 'test@example.com',
-      created_at: '2024-01-01T00:00:00Z',
-      updated_at: '2024-01-01T00:00:00Z'
+      items: [mockBankAccount],
+      total: 1
     });
   })
 ];
