@@ -19,14 +19,24 @@ export interface User {
   }
   
   // Bank Account types
-  export interface BankAccount {
-    id: string;
-    account_type: string;
-    account_name: string;
-    account_identifier: string;
-    created_at: string;
-    is_active: boolean;
-  }
+export interface BankAccount {
+  id: string;
+  user_id: string;
+  account_type: string;
+  account_name?: string;
+  account_identifier?: string;
+  created_at: string;
+  is_active: boolean;
+}
+
+export interface CreateBankAccountData {
+  account_type: string;
+  account_name?: string;
+  account_identifier?: string;
+  access_token: string;
+  refresh_token?: string;
+  token_expires_at?: string;
+}
   
   export interface PaginatedResponse<T> {
     items: T[];
@@ -47,4 +57,30 @@ export interface User {
   // Component prop types
   export interface LayoutProps {
     children?: React.ReactNode;
+  }
+
+  // Transaction types
+  export interface Transaction {
+    id: string;
+    user_id: string;
+    bank_account_id: string;
+    amount: number;
+    description?: string | null;
+    created_at: string;
+  }
+
+  export interface CreateTransactionData {
+    bank_account_id: string;
+    amount: number;
+    description?: string;
+  }
+
+  export interface UpdateTransactionData {
+    amount: number;
+    description?: string;
+  }
+
+  export interface TransactionList {
+    items: Transaction[];
+    total: number;
   }
